requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
python-whois>=0.9.4
dnspython>=2.6.1
lxml>=5.2.0
jinja2>=3.1.4
tabulate>=0.9.0
tqdm>=4.66.0
